In FSX, make the internal changed paths retrieval API streamy.
Instead of returning full lists, the new cached data layer API will
return only 1 block of changes, e.g. 100, and deliver more upon the
next call.  The size of a block is unspecified at the API level.
The current position within the changes list plus a number of other
context data is stored in a newly introduced context object.

This does not change the actual retrieval but above the cached data
layer, the code is now "final".  Also, the code path for in-txn data
will remain as-is because it needs to sanitize the whole changes list
before returning it and that can't be done iteratively.

* subversion/libsvn_fs_x/fs.h
  (svn_fs_x__changes_context_t): Define new internal API struct.

* subversion/libsvn_fs_x/cached_data.h
  (svn_fs_x__create_changes_context): Declare new internal API.
  (svn_fs_x__get_changes): Use two-pool paradigm now and take a
                           CONTEXT instead of FS+REV.

* subversion/libsvn_fs_x/cached_data.c
  (svn_fs_x__get_proplist): Implement.
  (svn_fs_x__get_changes): Take info from the CONTEXT now and update
                           the CONTEXT upon exit.

* subversion/libsvn_fs_x/tree.c
  (fs_revision_changes_iterator_data_t): Add the new retrieval context
                                         plus a temp pool.
  (x_revision_changes_iterator_get): Request data block by block.
  (x_report_changes): Update, performing the more complex query setup.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1733804 13f79535-47bb-0310-9956-ffa450edef68