Implement fancy descriptions for conflicts involving incoming edits.
Currently, this simply lists all revisions which changed the conflict
victim in the repository within the revision range of the operation.

We gather more details about the changes than are currently being displayed.
These details might be exposed in the description later, or we might expose
them via other (new) API interfaces. Not sure yet.

* subversion/libsvn_client/conflicts.c
  (conflict_tree_incoming_edit_details): New struct.
  (conflict_tree_get_details_incoming_edit): New. Type-specific get_details()
   implementation for 'incoming edit' conflicts.
  (conflict_tree_get_description_incoming_edit): New. Type-specific
   get_description() implementation for 'incoming edit' conflicts.
  (conflict_type_specific_setup): Add handlers for 'incoming edit'.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1735977 13f79535-47bb-0310-9956-ffa450edef68