Do not offer the 'ignore incoming deletion' option for real 'delete vs delete' tree conflicts. Offering such an option makes no sense in that case.
Since we now scan the log for incoming moves we can tell apart an incoming
deletion from an incoming move fairly reliably.

Arguably, an actual 'delete vs delete' conflict should be auto-resolved to
'accept incoming deletion', but that's left for later. The question of how
whether and how to provide default answers for conflict resolver questions
should be settled first.

* subversion/libsvn_client/conflicts.c
  (configure_option_incoming_delete_ignore): Don't add this option in a real
   'delete vs delete' conflict situation.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1751914 13f79535-47bb-0310-9956-ffa450edef68