Fix a bug in mod_dav_svn's MERGE handler that I introduced in r1758224.
The problem is caused by passing a wrong request argument to the new
dav_svn__output_create() function.  Within the merge handler, the are
two dav_resource instances — the target and the source, with the second
being created via a subrequest.  We were incorrectly using the request of
the source resource, and writing the response to the subrequest's output
filter stack.  Doing so can result in malformed responses.

This bug can be triggered by committing with Apache Httpd 2.2.x and
mod_deflate (or by running the tests with the new --mod-deflate switch):

  Committing transaction...
  svn: E175009: Commit failed (details follow):
  svn: E175009: The XML response contains invalid XML
  svn: E130003: Malformed XML: not well-formed (invalid token)

* subversion/mod_dav_svn/version.c
  (merge): Pass the request for the target resource when calling
   dav_svn__output_create().


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1758781 13f79535-47bb-0310-9956-ffa450edef68