Add a workaround for yet another issue with APR's apr_file_trunc.
The previous workaround is ineffective if the last file access had been
a read.  Now, we force it into to "write mode" internally to have the
existing workaround kick in.

Luckily, this only affects 'svnadmin pack' for FSFS format 7 and FSX.
The other functions using trunc should have no problem with the added
overhead.

* subversion/libsvn_subr/io.c
  (svn_io_file_trunc): Admend the existing workaround with a dummy-write.

* subversion/tests/libsvn_subr/io-test.c
  (test_apr_trunc_workaround): New test demonstrating the problem.
  (test_funcs): Register the new test.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1759116 13f79535-47bb-0310-9956-ffa450edef68