Change the way the conflict resolver merges incoming moves.
When ^/trunk/A is moved to ^/trunk/B and this move is merged to a working
copy of ^/branch, we'd like to create a corresponding move from ^/branch/A
to ^/branch/B. However, the merge adds the copy-half of the incoming move
as a copy from ^/trunk/B.

The resolver used to merge changes from ^/branch into B (copied from ^/trunk/B),
and then performed a meta-data-only move from A to B. This approach has some
problems, such as unwanted 'replacements' of children.
For example:
If the branch added file A/new, trunk moved A to B, and we're merging trunk
into a wc of branch, then the 'merge incoming move' option would produce a
merge commit where B/new was replaced:
  D /branch/A
  A /branch/B (from /branch/A:4)
  R /branch/B/new (from /branch/A/new:4)
This replacment happened because the meta-data-only move of A to B created
a new undesired op-depth layer in wc.db.

So we'll try another approach: Revert the copy of B added by the merge,
move A to B locally, and then run a merge from ^/trunk into B.
This is not perfect either but the end result is closer to what we would like.
Reverting local changes might not always be a good idea but the resolver
already assumes that users do not make manual changes to the working copy
before resolving conflicts. And reverting an otherwise unmodified copy of
an existing repository subtree cannot lead to data loss.

* subversion/libsvn_client/conflicts.c
  (resolve_incoming_move_dir_merge): Use a "revert, then move, then merge"
   resolution strategy as described above.

* subversion/tests/libsvn_client/conflicts-test.c
  (test_merge_incoming_move_dir,
   test_merge_incoming_move_dir2): Fix test expectations.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1761029 13f79535-47bb-0310-9956-ffa450edef68