In the tree conflict resolver, remove the options that simultaneosly resolve the tree *and* the potential text conflict in the file "add vs add" case.
This turns the following set of options ...

    (i) ignore incoming addition, (m) merge the files,
    (R) delete my file and replace it with incoming file,
    (M) replace my file with incoming file and merge the files,

...into this:

    (m) merge the files,
    (M) replace my file with incoming file and merge the files,

The idea behind this change is that it would be better if the tree conflict
resolver would only handle the "tree conflict" part, without doing anything
about a potential text conflict.  So, we only leave the options to pick the
proper line of history for the file (either by keeping the local line of
history or by replacing it with the incoming one), but always merge the
files.  If the merge provokes a text conflict, it could then be resolved
separately.

Another reason is that, by cutting down the amount of options to 2, we have
a possibility to add totally different options, like "move my locally added
file to foo.1".  With 4 existing options, as it was before this change, it
would be impossible, because 6 non-default options is just too much for the
user.

(I kept the analogues of these options for directories for now.)

* subversion/include/svn_client.h
  (svn_client_conflict_option_id_t): Remove
   svn_client_conflict_option_incoming_added_file_replace.

* subversion/libsvn_client/conflicts.c
  (verify_local_state_for_incoming_add_upon_update): Don't test for the
   option id which is now gone.
  (configure_option_incoming_add_ignore): Tweak this configuration function
   to only add the option for directories.
  (configure_option_incoming_added_file_replace): Remove.
  (svn_client_conflict_tree_get_resolution_options): Don't call the now
   removed configuration function.

* subversion/svn/conflict-callbacks.c
  (builtin_resolver_options): Remove entry that corresponds to deleted
   svn_client_conflict_option_incoming_added_file_replace.

* subversion/tests/libsvn_client/conflicts-test.c
  (test_merge_incoming_added_file_ignore,
   test_merge_incoming_added_file_replace,
   test_update_incoming_added_file_ignore,
   test_update_incoming_added_file_replace,
   test_switch_incoming_added_file_ignore): Remove these tests.
  (test_funcs): Update the test list.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1764435 13f79535-47bb-0310-9956-ffa450edef68