In the tree conflict resolver, fix merging text content in the "remote move" vs "local edit" scenario.
The issue was that if the conflict happened during merge, we were
using mismatched right and target files when calling svn_wc_merge5().
For instance, if we're merging from trunk to branch, we would want

  - "left" file to have the ancestor's content
  - "right" file to have the content from trunk
  - "target" file to have the content from branch

Since we were first doing a merge on the added part of the move, and then
doing the metadata-only move, the "target" file (that arrived during merge)
was actually having content from trunk, and the "right" file had content
from branch.

Fix this by putting aside the target's file content, and moving the file
*before* doing a three-way merge.  Ideally, this should be performed on
the libsvn_wc layer, and should be using a workqueue, but this is something
left for future work.

* subversion/libsvn_client/conflicts.c
  (resolve_incoming_move_file_text_merge): Check the conflict operation
   at the beginning of this function.  If the conflict happened as a result
   of the merge, set aside the move target file, perform the move, and then
   do a proper three-way merge.

* subversion/tests/libsvn_client/conflicts-test.c
  (test_merge_incoming_move_file_text_merge_conflict): Remove XFAIL comment.
  (test_funcs): The test_merge_incoming_move_file_text_merge_conflict() test
   now passes.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1765466 13f79535-47bb-0310-9956-ffa450edef68