Fix a long-standing FSFS bug that causes read errors on valid / non-corrupted representations.  The problem occurs if
    * the representation is a DELTA against a PLAIN one,
    * spans multiple delta windows, and
    * has a matching sub-string in some window after there were no matches
      in the previous window.

Because we used to read the PLAIN base window only if it was really needed
due to a matching sub-string In that case, the read offset inside the PLAIN
rep would not be updated.  As a result, when the next window needs to be
read as the delta base, the previous one is read instead.  This would almost
certainly result in a corrupted reconstructed fulltext.  The read will
ultimately fail due to a checksum mismatch.

A test case and issue tracker entry will follow.

* subversion/libsvn_fs_fs/cached_data.c
  (skip_plain_window): New function.
  (get_combined_window): Always move the offset within the base rep in sync
                         with the delta rep.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1766352 13f79535-47bb-0310-9956-ffa450edef68