Second attempt to preserve local changes for files which become victims of a nested incoming move during an update (same problem as in r1771924).
This time, we make sure the file is copied to the move destination before
the move source is deleted. The file becomes the victim of a new tree conflict
at the move destination, which the resolver could resolve later on.
Information about the file's original location is preserved in conflict data.

* subversion/libsvn_wc/wc_db_update_move.c
  (tc_editor_update_incoming_moved_file): If the file cannot be found at the
   move destination, copy it there and raise a new tree conflict.

* subversion/tests/libsvn_client/conflicts-test.c
  (test_update_incoming_dir_move_with_nested_file_move): Update test
   expectations accordingly. And verify that expected file content is present.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1772053 13f79535-47bb-0310-9956-ffa450edef68