In the test suite, provide and use a sandbox-specific method when we need to create a local config directory.
This makes the path and name of the config directory more consistent, and
enforces segregation of directories used by different tests. Not that any of
the existing tests did use clashing directory paths, but some of them were a
bit "too close for comfort". Using this will ensure they can't in future. It
also simplifies the call site.

This is just something I found that could be improved -- there was no bug
report or issue related to it.

* subversion/tests/cmdline/autoprop_tests.py
  (create_config,
   create_inherited_autoprops_config): Use sbox.create_config_dir().
  (autoprops_test,
   inheritable_autoprops_test,
   svn_prop_inheritable_autoprops_add_versioned_target): Adjust their callers.

* subversion/tests/cmdline/basic_tests.py
  (basic_auth_test): Use sbox.create_config_dir().

* subversion/tests/cmdline/import_tests.py
  (import_inherited_ignores): Adjust call to create_inherited_ignores_config().

* subversion/tests/cmdline/prop_tests.py
  (create_inherited_ignores_config): Use sbox.create_config_dir().
  (inheritable_ignores): Adjust its caller.

* subversion/tests/cmdline/stat_tests.py
  (timestamp_behaviour): Use sbox.create_config_dir().

* subversion/tests/cmdline/svntest/sandbox.py
  (Sandbox.create_config_dir): New method.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1773439 13f79535-47bb-0310-9956-ffa450edef68