Tweak the way the conflict resolver discovers nested moves.
This is no functional change. The effective change is that the resolver now
builds an internal list of moves which includes all nested moves found in
each revision. Previously it was adding just one particular nested move to
the list of moves, a nested move which was deemed to apply to the conflict
victim. So we now discover the full list which includes all nested moves,
and then search that list.

* subversion/libsvn_client/conflicts.c
  (map_deleted_path_to_move): Make this function recurse. If possible this
   function now finds the effective move of a node which was moved along
   with its parent and then moved again (in the same revision).
  (find_nested_move): Rename to ...
  (find_nested_moves): .. this, and return a list of all nested moves found
   in the revision, instead of perhaps returning just one nested move which
   applies to the conflict victim.
  (find_deleted_rev): Build a full list of moves in the current revision,
   including all nested ones. Then look the deleted node up in this list.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1775111 13f79535-47bb-0310-9956-ffa450edef68