Reduce memory usage of merge by using a subpool for temporary mergeinfo.
In my test case this reduced the total memory usage (RSS on Linux) of the
process e.g. from baseline 8 MB + growth of 50 MB to baseline 8 MB + growth
of 43 MB.

My test case created 300 branches, each with a small change to a different
file, merged all of them to trunk in such a way that trunk ended up with a
lot of subtree mergeinfo (300 files with mergeinfo, 300 branches mentioned
in each one's mergeinfo), and then monitored the memory usage of an
automatic merge from trunk to the first branch.

* subversion/libsvn_client/merge.c
  (get_mergeinfo_paths): Use a subpool for temporary mergeinfo.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1776788 13f79535-47bb-0310-9956-ffa450edef68