Eliminate a server roundtrip from checkouts of HEAD.
Prior to the patch, the head reavision would be requested twice; once to
resolve the PATHREV and a second time in the update_internal call.  This
is not only unnecessary but also yields inconsistent results as we use the
first result to check the patch to check out while the second may be a
later revision where the path has already been deleted.

* subversion/libsvn_client/checkout.c
  (svn_client__checkout_internal):  We have resolved the resivision number
                                    already and checked the path's type for
                                    it.  So, run checkout for exactly the
                                    revision we just checked.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1779620 13f79535-47bb-0310-9956-ffa450edef68