Fix a hole in FSFS's checksum usage.
If we feed a file content stream from the fulltext cache, FSFS won't
calculate a checksum over it because we already checked it when reading
it the first time.  However, if the file content does get evicted from
the cache while the stream has not been fullt read, yet, the remainder
must come from reconstructed text windows.  The current code did not
compare the checksum in this case - leaving it to the ra/client side
to detect intermediate corruptions.

This patch ensures we calculate the checksum over the full file content.
It then gets automatically verified once the last byte is being delivered.

Found by: julianfoad

* subversion/libsvn_fs_fs/cached_data.c
  (skip_contents): Run the checksum over the skipped part as well, so
                   that OFF will eventually match the fulltext length.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1781694 13f79535-47bb-0310-9956-ffa450edef68